[FE] fix: 필요없는 파일 삭제 (#113)
