+++ fe/src/components/issueDetailPage/Body.tsx
@@ -1,7 +1,14 @@
+import { useTheme } from '@emotion/react';
 import { Theme, css } from '@emotion/react';
-import { SideBarRightPanel } from './SideBarRightPanel';
+import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
 import { CommentContainer } from './CommentContainer';
 import { Comment } from '@components/common/comment/Comment';
+import { ListSideBar } from '@components/common/sideBar/ListSideBar';
+import { SideBar } from '@components/common/sideBar/SideBar';
+import { ReactComponent as Trash } from '@assets/icons/trash.svg';
+import { Button } from '@components/common/Button';
+import { deleteIssue } from 'apis/api';
 
 type Props = {
   issueDetailPageData: IssueDetailPageData;
@@ -26,6 +33,21 @@ export const Body: React.FC<Props> = ({
   onMultipleSelectedAssignee,
   onMultipleSelectedLabel,
 }) => {
+  const theme = useTheme() as any;
+  const [isDeleteError, setIsDeleteError] = useState<boolean>(false);
+  const navigate = useNavigate();
+
+  const onDeleteIssue = async () => {
+    try {
+      setIsDeleteError(false);
+      await deleteIssue(issueDetailPageData.id);
+      navigate('/issues');
+    } catch (error) {
+      console.error(error);
+      setIsDeleteError(true);
+    }
+  };
+
   return (
     <div css={bodyStyle}>
       <div className="container">
@@ -42,15 +64,27 @@ export const Body: React.FC<Props> = ({
           onDeleteComment={onDeleteComment}
         />
       </div>
-      <SideBarRightPanel
-        issueId={issueDetailPageData.id}
-        selectionsOptions={selectionsOptions}
-        selections={selections}
-        onChangeSelect={onChangeSelect}
-        onSingleSelectedMilestone={onSingleSelectedMilestone}
-        onMultipleSelectedAssignee={onMultipleSelectedAssignee}
-        onMultipleSelectedLabel={onMultipleSelectedLabel}
-      />
+      <div className="rightPanel">
+        <SideBar>
+          <ListSideBar
+            onSingleSelectedMilestone={onSingleSelectedMilestone}
+            onMultipleSelectedAssignee={onMultipleSelectedAssignee}
+            onMultipleSelectedLabel={onMultipleSelectedLabel}
+            selectionsOptions={selectionsOptions}
+            selections={selections}
+            onChangeSelect={onChangeSelect}
+          />
+        </SideBar>
+        <Button typeVariant="ghost" size="S" onClick={onDeleteIssue}>
+          <Trash stroke={theme.danger.text.default} />
+          <span className="button-delete">이슈 삭제</span>
+        </Button>
+        {isDeleteError && (
+          <span css={{ color: theme.danger.text.default }}>
+            이슈 삭제에 실패했습니다.
+          </span>
+        )}
+      </div>
     </div>
   );
 };
@@ -67,5 +101,19 @@ const bodyStyle = (theme: Theme) => css`
     display: flex;
     flex-direction: column;
     gap: 24px;
+  }
+
+  .rightPanel {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    align-items: flex-end;
+    & span {
+      color: ${theme.neutral.text.default};
+      font: ${theme.fonts.availableMedium16};
+    }
+    .button-delete {
+      color: ${theme.danger.text.default};
+    }
   }
 `;
